refactor(datahub): migrate CalendarStore to domains/metadata/calendar/
迁移内容:
- 创建 CalendarDay dataclass (models.py)
- 迁移 CalendarStore 完整实现 (calendar_store.py)
- 创建 __init__.py 导出模块
- 在旧位置添加弃用警告和重新导出

更新导入路径:
- packages/datahub/src/ditto_datahub/accessors/calendar_accessor.py
- packages/datahub/src/ditto_datahub/runtime/sql_engine.py
- apps/port/src/ditto_port/registry/datahub.py
- 所有测试文件

验证:
- 所有单元测试通过 (25 passed)
- 所有集成测试通过 (3 passed)
- 类型检查通过 (0 errors)
- 向后兼容性保持 (弃用警告)

遵循与 SecurityStore 相同的迁移模式。
