+++ packages/datahub/tests/integration/runtime/test_sql_engine_integration.py
@@ -4,8 +4,8 @@ from pathlib import Path
 from tempfile import TemporaryDirectory
 
 import polars as pl
+from ditto_datahub.domains.metadata.calendar import CalendarStore
 from ditto_datahub.runtime.sql_engine import SqlEngine
-from ditto_datahub.stores.calendar_store import CalendarStore
 from ditto_datahub.stores.security_store import SecurityStore
 from ditto_datahub.stores.sqlite_client import SQLiteClient
 from ditto_foundation import SQLitePool
