+++ packages/datahub/tests/unit/accessors/test_calendar_accessor_unit.py
@@ -2,7 +2,7 @@
 
 import pytest
 from ditto_datahub.accessors.calendar_accessor import CalendarAccessor
-from ditto_datahub.stores.calendar_store import CalendarStore
+from ditto_datahub.domains.metadata.calendar import CalendarStore
 from ditto_datahub.stores.sqlite_client import SQLiteClient
